INT-2130 - Refactored to move scripting support from core to scripting
