+++ spring-integration-groovy/src/main/resources/META-INF/spring.schemas
