+++ spring-integration-scripting/src/main/java/org/springframework/integration/scripting/ScriptVariableGenerator.java
