+++ spring-integration-scripting/src/main/resources/META-INF/spring.handlers
@@ -1,1 +1,1 @@
-http\://www.springframework.org/schema/integration/script=org.springframework.integration.scripting.config.jsr223.ScriptNamespaceHandler
+http\://www.springframework.org/schema/integration/scripting=org.springframework.integration.scripting.config.jsr223.ScriptNamespaceHandler
