+++ spring-integration-scripting/src/main/resources/META-INF/spring.schemas
@@ -1,2 +1,4 @@
-http\://www.springframework.org/schema/integration/script-2.1.xsd=org/springframework/integration/script/config/spring-integration-script-2.1.xsd
-http\://www.springframework.org/schema/integration/script.xsd=org/springframework/integration/script/config/spring-integration-script-2.1.xsd
+http\://www.springframework.org/schema/integration/scripting-2.1.xsd=org/springframework/integration/scripting/config/spring-integration-scripting-2.1.xsd
+http\://www.springframework.org/schema/integration/scripting.xsd=org/springframework/integration/scripting/config/spring-integration-scripting-2.1.xsd
+http\://www.springframework.org/schema/integration/scripting-core-2.1.xsd=org/springframework/integration/scripting/config/spring-integration-scripting-core-2.1.xsd
+http\://www.springframework.org/schema/integration/scripting-core.xsd=org/springframework/integration/scripting/config/spring-integration-scripting-core-2.1.xsd
