+++ spring-integration-scripting/src/main/resources/META-INF/spring.tooling
@@ -1,4 +1,4 @@
-# Tooling related information for the integration script namespace
-http\://www.springframework.org/schema/integration/script@name=integration script Namespace
-http\://www.springframework.org/schema/integration/script@prefix=int-script
-http\://www.springframework.org/schema/integration/script@icon=org/springframework/integration/script/config/spring-integration-script.gif
+# Tooling related information for the integration scripting namespace
+http\://www.springframework.org/schema/integration/scripting@name=integration scripting Namespace
+http\://www.springframework.org/schema/integration/scripting@prefix=int-script
+http\://www.springframework.org/schema/integration/scripting@icon=org/springframework/integration/scripting/config/spring-integration-scripting.gif
