+++ spring-integration-scripting/src/main/resources/org/springframework/integration/scripting/config/spring-integration-scripting-core-2.1.xsd
