+++ spring-integration-scripting/src/main/resources/org/springframework/integration/scripting/config/spring-integration-scripting.gif
