+++ spring-integration-scripting/src/test/java/org/springframework/integration/scripting/config/jsr223/Jsr223FilterTests-context.xml
@@ -2,10 +2,10 @@
 <beans:beans xmlns="http://www.springframework.org/schema/integration"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xmlns:beans="http://www.springframework.org/schema/beans"
-	xmlns:script="http://www.springframework.org/schema/integration/script"
+	xmlns:script="http://www.springframework.org/schema/integration/scripting"
 	xsi:schemaLocation="http://www.springframework.org/schema/integration http://www.springframework.org/schema/integration/spring-integration.xsd
 		http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-		http://www.springframework.org/schema/integration/script http://www.springframework.org/schema/integration/script.xsd">
+		http://www.springframework.org/schema/integration/scripting http://www.springframework.org/schema/integration/scripting-2.1.xsd">
 
 	<filter input-channel="referencedScriptInput">
 		<script:script lang="groovy" location="org/springframework/integration/scripting/config/jsr223/Jsr223FilterTests.groovy"/>
